lib/size: GetSize returns uint64
As per spec:
https://docs.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-getdiskfreespaceexw

+++ lib/size_linux.go
@@ -7,7 +7,7 @@ import (
 )
 
 // GetSpace returns disk size information for the given drive (bytes).
-func GetSpace(drive string) (free int64, total int64, avail int64, err error) {
+func GetSpace(drive string) (free uint64, total uint64, avail uint64, err error) {
 	var stat unix.Statfs_t
 	err = unix.Statfs(drive, &stat)
 	if err != nil {
@@ -17,9 +17,9 @@ func GetSpace(drive string) (free int64, total int64, avail int64, err error) {
 	// blocks * size per block = available space in bytes
 	size := uint64(stat.Bsize)
 
-	total = int64(stat.Blocks * size)
-	avail = int64(stat.Bavail * size)
-	free = int64(stat.Bfree * size)
+	total = stat.Blocks * size
+	avail = stat.Bavail * size
+	free = stat.Bfree * size
 
 	return avail, total, free, nil
 }
